refactor(extract): converts svelte wrapper to esm
