+++ package.json
@@ -170,7 +170,7 @@
     "rechoir": "^0.8.0",
     "safe-regex": "2.1.1",
     "semver": "^7.3.7",
-    "semver-try-require": "6.2.1",
+    "semver-try-require": "6.2.2",
     "teamcity-service-messages": "0.1.14",
     "tsconfig-paths-webpack-plugin": "4.0.1",
     "watskeburt": "0.10.1",
