+++ test/extract/transpile/meta.spec.mjs
@@ -75,8 +75,10 @@ describe("[U] extract/transpile/meta", () => {
     expect(
       getWrapper(".svelte", {
         babelConfig: { babelrc: false },
-      }).transpile.toString()
-    ).to.contains("svelte");
+      })
+        .transpile("")
+        .toString()
+    ).to.contain("generated by Svelte");
   });
 
   it("returns me the available transpilers", () => {
