+++ test/extract/transpile/svelte-wrap.spec.mjs
@@ -1,6 +1,6 @@
-import { promises as fsPromise } from "node:fs";
+import { readFileSync } from "node:fs";
 import { expect } from "chai";
-import svelteWrap from "../../../src/extract/transpile/svelte-wrap.js";
+import svelteWrap from "../../../src/extract/transpile/svelte-wrap.mjs";
 import normalizeSource from "../normalize-source.utl.mjs";
 import typescriptWrap from "../../../src/extract/transpile/typescript-wrap.mjs";
 
@@ -21,24 +21,19 @@ describe("[I] svelte transpiler", () => {
         ),
     ],
   ].forEach(([variant, transformExpected]) => {
-    it(`'transpiles' svelte with "<script lang='${variant}'>"'`, async () => {
-      const observedPromise = fsPromise
-        .readFile(
-          `./test/extract/transpile/__mocks__/svelte-${variant}.svelte`,
-          "utf8"
-        )
-        .then((pSourceCode) => wrap.transpile(pSourceCode));
-      const expectedPromise = fsPromise.readFile(
+    it(`'transpiles' svelte with "<script lang='${variant}'>"'`, () => {
+      const lSource = readFileSync(
+        `./test/extract/transpile/__mocks__/svelte-${variant}.svelte`,
+        "utf8"
+      );
+      const lObserved = wrap.transpile(lSource);
+      const lExpected = readFileSync(
         "./test/extract/transpile/__fixtures__/svelte.js",
         "utf8"
       );
 
-      const [observed, expected] = await Promise.all([
-        observedPromise,
-        expectedPromise,
-      ]);
-      expect(normalizeSource(observed)).to.equal(
-        normalizeSource(transformExpected(expected))
+      expect(normalizeSource(lObserved)).to.equal(
+        normalizeSource(transformExpected(lExpected))
       );
     });
   });
